# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "开源、自托管轻量级发布平台 Ech0 的 API 文档",
        "title": "Ech0 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6277",
    "basePath": "/api",
    "paths": {
        "/access-tokens": {
            "get": {
                "description": "列出当前用户的所有访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "列出访问令牌",
                "responses": {
                    "200": {
                        "description": "列出访问令牌失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "为当前用户创建一个新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "创建访问令牌",
                "parameters": [
                    {
                        "description": "新的访问令牌信息",
                        "name": "accessToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AccessTokenSettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建访问令牌失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/access-tokens/{id}": {
            "delete": {
                "description": "根据 ID 删除指定的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "删除访问令牌",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要删除的访问令牌 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除访问令牌失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/agent/settings": {
            "get": {
                "description": "获取系统的 Agent 相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取 Agent 设置",
                "responses": {
                    "200": {
                        "description": "获取 Agent 设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新系统的 Agent 相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新 Agent 设置",
                "parameters": [
                    {
                        "description": "新的 Agent 设置",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AgentSettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新 Agent 设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/backup/schedule": {
            "get": {
                "description": "获取系统的定期备份计划设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取备份计划",
                "responses": {
                    "200": {
                        "description": "获取备份计划失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "为系统设置定期备份计划",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新备份计划",
                "parameters": [
                    {
                        "description": "备份计划设置",
                        "name": "backupSchedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackupScheduleDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置备份计划失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/comment/settings": {
            "get": {
                "description": "获取系统的评论相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取评论设置",
                "responses": {
                    "200": {
                        "description": "获取评论设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新系统的评论相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新评论设置",
                "parameters": [
                    {
                        "description": "新的评论设置",
                        "name": "commentSettings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CommentSettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新评论设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/oauth2/settings": {
            "get": {
                "description": "获取系统的 OAuth2 相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取 OAuth2 设置",
                "responses": {
                    "200": {
                        "description": "获取 OAuth2 设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新系统的 OAuth2 相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新 OAuth2 设置",
                "parameters": [
                    {
                        "description": "新的 OAuth 设置",
                        "name": "oauthSettings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OAuth2SettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新 OAuth 设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/s3/settings": {
            "get": {
                "description": "获取系统的 S3 存储相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取 S3 存储设置",
                "responses": {
                    "200": {
                        "description": "获取 S3 存储设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新系统的 S3 存储相关设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新 S3 存储设置",
                "parameters": [
                    {
                        "description": "新的 S3 存储设置",
                        "name": "s3Settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.S3SettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新 S3 存储设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/settings": {
            "get": {
                "description": "获取系统的全局设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取设置",
                "responses": {
                    "200": {
                        "description": "获取设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新系统的全局设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "更新设置",
                "parameters": [
                    {
                        "description": "新的系统设置",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SystemSettingDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新设置失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据，具体内容因接口而异"
                },
                "msg": {
                    "description": "Msg 返回信息，通常是状态描述",
                    "type": "string"
                }
            }
        },
...
        "model.AccessTokenSettingDto": {
            "type": "object",
            "properties": {
                "expiry": {
                    "description": "访问令牌过期策略（8_hours/1_month/never）",
                    "type": "string"
                },
                "name": {
                    "description": "访问令牌名称",
                    "type": "string"
                }
            }
        },
...
        "model.AgentSettingDto": {
            "type": "object",
            "properties": {
                "api_key": {
                    "description": "LLM API Key",
                    "type": "string"
                },
                "base_url": {
                    "description": "自定义 API URL（可选）",
                    "type": "string"
                },
                "enable": {
                    "description": "是否启用 Agent 功能",
                    "type": "boolean"
                },
                "model": {
                    "description": "LLM 模型名称",
                    "type": "string"
                },
                "prompt": {
                    "description": "Agent 额外使用的提示词",
                    "type": "string"
                },
                "provider": {
                    "description": "LLM 提供商 （OpenAI、DeepSeek、Anthropic、Gemini、阿里百炼、Ollama等）",
                    "type": "string"
                }
            }
        },
...
        "model.BackupScheduleDto": {
            "type": "object",
            "properties": {
                "cron_expression": {
                    "description": "备份计划的 Cron 表达式",
                    "type": "string"
                },
                "enable": {
                    "description": "是否启用备份计划",
                    "type": "boolean"
                }
            }
        },
...
        "model.CommentSettingDto": {
            "type": "object",
            "properties": {
                "comment_api": {
                    "description": "评论 API 地址",
                    "type": "string"
                },
                "enable_comment": {
                    "description": "是否启用评论",
                    "type": "boolean"
                },
                "provider": {
                    "description": "评论提供者",
                    "type": "string"
                }
            }
        },
...
    }
}